Add GetCurrentForecastUseCase

+++ app/src/main/java/com/bboykot/weather/domain/models/CurrentForecast.kt
@@ -2,7 +2,7 @@ package com.bboykot.weather.domain.models
 
 data class CurrentForecast(
     val city: String,
-    val temperature: Int,
+    val temperature: Double,
     val wind: Double,
     val description: String,
 )
